Defer measurement notification to new threads - a heavyweight solution.
This avoids making the data source block waiting to return from the callback.
This commit also adds a lock in each thread to make sure only one notification
is going out to the callbacks at a time, since the RemoteCallackList doesn't
allow this.

A better solution will be to dump the updates into a queue which has only one
long-lived thread to process it and make the notifications.