Avoid OOMing the machine if you run Eclipse and try to install this.
I bumped this up for Mac OS X which had a really low max and wasn't allowing the
dx compile to finish, but this level is causing my 4GB machine to start swapping
and become unresponsive if I also have DDMS running in Eclipse. We'll try this,
but it may just lead to problems in OS X again.