Add method to wait for VehicleService to be ready to go.
If you really truly must wait until the vehicle service is able to start giving
you measurements, you can call the new waitUntilBound() method to block until
the remote vehicle service is initialized and connected. This could never return
if there's an error with the remote service.

Most applications will not need this, but it's useful for testing.