Listen for USB attachment intents in a proxy Activity.
This does receive the Intent, but the act of opening up either the activity or
the pop-up to confirm that the activity should start pauses the underlying app
(e.g. VehicleDashboard) and thus the VehicleService and RemoteVehicleService
lose their bound client and exit. So when this activity sends our custom
"refresh devices" intent, the UsbVehicleDataSource isn't around to actually
receive it.