Don't permaturely exit the USB data source - seems to fix race condition.
The NotificationTherad seems to always get the InterruptedException now, so
maybe poll() is unneccessary and we can go back to take. My reasoning of the
existence of a race condition seems sound, but perhaps it wasn't ever ocurring.
I'll leave it for now since it does seem more correct, even if it has a very
tiny bit of overhad (waking up every 1s to poll).