Don't throw NoValueException from Measurement.getValue().
This is an attempt to minimize the amount of boilerplate code required to access
a single measurement from the vehicle. Previously, every call to
Measurement.getValue() required wrapping in a try/catch block for the
NoValueException. This is because Measurement inherited from NoneData which
enabled measurements to exist with no value. You could programatically check if
they had a value or not before using them.

The motivation for this is that sometimes when the system isn't ready yet, it
can't give you a value for a requested measurement. In an effort to not throw
spurious errors, it was returning a blank measurement.

However, the common use case for OpenXC is asynchronous, and the system doesn't
really need to call your measurement listener if it knows the measurement has no
value. Therefore the "no value" error only really needs to happen when you
synchronously query for a measurement and we can't return one.

This commit moves the NoValueException from Measurement.getValue() to
VehicleService.get(), and removes the NoneData class.