Marshall measurement values to doubles in RawMeasurement, not the callback.
This is a pretty big but logical change - the marshalling to doubles really had
nothing to do with the data source or the measurements, it's just an odditiy of
the RemoteVehicleService situation.

This refactoring allowed me to do the data sink dumping at an earlier point,
before we've lost the true value of the measurements coming over USB (e.g. true
instead of 1).