Move all tests and dependencies to the openxc-it package, separate from the lib.
There are a number of bugs in the Maven/Android/Eclipse integration that cause
transitive dependencies to be included in the packaged APK file a number of
times, causing errors. By moving all of the test code (which includes some
packages that I found to have this transititive dependency issue, like junit and
hamcrest), the openxc library should be clear to import by anyone in Eclipse.