Jot down some initial thoughts on how the write API fits into the system.
Things are getting a little bit weird because we assume that we stop and start
sources whenever we want them to pump data into the system, but in this case we
need to keep the one USB connection alive. I added a new close() method to the
USB device class that does the true shutdown, but I need to think about it more.