Search for the USB device every time the USB thread starts.
This fixes a problem where if the USB source is start()ed, then stop()ed, then
started() again -- all while the USB device is plugged in -- we never try to
connect to it again.

The old workflow was this:

* When the USB source is created, initialize a connection to the device. Also
    start listening for unplug/plug events so we can detect a USB device being
    plugged in, and start up the USB thread when we detect that.

Of course, what if the source is restarted (but without calling the constructor
again), and the USB device remains plugged in? We don't try to connect.

The fix is to try and initialize the device when the thread starts, not when the
USB source object is created.