Use common byte stream handling code for Bluetooth and USB.
Because we have a socket stream with Bluetooth, we can use BufferedReader and
the nice readLine method. I tried implementing this for USB in the
bufferedreader branch with a custom InputStream, but the synchronization
required in my first attempt really killed performance.

In an effort to have the data sources use the same code, I've reworked the
Bluetooth source to use our custom newline-separated message parser in
BytestreamDataSourceMixin.