Break out of a possible infinite loop with invalid JSON.
I think this may have been the root of many odd Bluetooth problems. In
some cases, we could load a partial message as JSON but it would have
corrupted field names, and at some point the "field" in this function
would be null and we would loop forever. I noticed this because when
testing the foreground service notification, I was not getting the
"disconnected" callback from the BT data source if I unplugged the VI
abruptly.