Kick off BT connection attempts when another BT device connects.
If you're getting into a car, your car is probably connecting to the
phone/audio socket on your phone - it's a good indicator that we should
re-scan (if you have polling enabled or BT devices). If you have a new
VI that can act as master and have polling disabled, this will have no
effect.

Fixed #121.