Don't re-connect BT if we are already making a connection.
The problem is that when we detect a new device, we were restarting the
connection attempt task, but in the process we would lose reference to
an existing thread that was making connection attempts.

1. The first thread is blocked waiting for a socket connection attempt to
fail (it started too soon, before the device was available)o
2. The device is discovered or connected, and we kick off another
connection thread.
3. The new connection thread immediately succeeds and you get an active
connection. Yay!
4. The old thread fails and since it assumes it's the only thing around,
it kills the connection we just made.

The fix is pretty simple - don't reset connection attempts unless we've
already given up.

Fixed #126.