Catch a NoClassDefFound error when compiling in Eclipse with Java 1.7.
I'm not sure what's up, but when building the Enabler from Eclipse we
get a NoClassDefFoundError when registering the Bugsnag token. From
Maven, it works fine. I think this has to do with JDK 7 support, since
the Bugsnag library is compiled with 1.6. However, even if I rewrite the
project so it again compiles with JDK 6 I still aren't able to get this
to work. For now, we just catch the error here and continue. The release
builds are done with Maven.