Change Android package namespace to com.openxcplatform.com
I'm trying to deploy to the Google Play store, but someone has already
uploaded the Enabler using our original package name. I can't seem to
find it in the store to attempt to ask this person to give it up to the
platform maintainers. It's OK since we should be using the full domain
name for the package anyway.