Add map pruning typically handled automatically by Multimap.
With a Multimap, removing a key-value pair, where that value is the last one for the key, removes that key. Switching to a map instead of a Multimap, no such 'automatic' functionality exists because the MessageListenerGroup instance is still associated with that key, even if it contains no listeners, so you have to tell the map to delete it manually.