Don't remove VI constructors used via reflection with ProGuard.
Fixed #236.

It turns out that ProGuard (which is only used on the release build and
not the debug builds that I typically test with), was being too
aggressive. It stripped out the constructors for some of the
VehicleInterface classes, so we weren't able to create instances of
those clases. Because ProGuard could never find a time when those
constructors were used, it stripped them out. When the factory tried to
create a new BluetoothVehicleInterface instance, it would get that
error.

The constructors are never used explicitly because they are constructed
using reflection at runtime in the VehicleInterfaceFactory. The reason
for this is so that the Enaber can send a message to the background
processing running the VI connection threads, and it doesn't require a
hard coded conditional to find the right class for the VI.

This updates the ProGuard configuration to not strip out any public or
protected functions in the com.openxc package - it's all open source
code, so we don't need to be concerned with obfuscation, and the
codebase is small enough to not need to worry about optimization too
much. I confirmed that with these classes explicitly kept in the
ProGuard config, it connects just fine.